Reading of LNP Header v4 - OK
